+++ lib/features/credit_card/presentation/ui/credit_card_form_page.dart
@@ -6,6 +6,7 @@ import 'package:finan_master_app/features/account/presentation/ui/components/acc
 import 'package:finan_master_app/features/credit_card/domain/entities/credit_card_entity.dart';
 import 'package:finan_master_app/features/credit_card/domain/enums/brand_card_enum.dart';
 import 'package:finan_master_app/features/credit_card/presentation/notifiers/credit_card_notifier.dart';
+import 'package:finan_master_app/features/credit_card/presentation/states/credit_card_state.dart';
 import 'package:finan_master_app/features/credit_card/presentation/ui/components/card_brand_list_bottom_sheet.dart';
 import 'package:finan_master_app/shared/classes/form_result_navigation.dart';
 import 'package:finan_master_app/shared/extensions/double_extension.dart';
@@ -66,134 +67,139 @@ class _CreditCardFormPageState extends State<CreditCardFormPage> with ThemeConte
   @override
   Widget build(BuildContext context) {
     return ValueListenableBuilder(
-        valueListenable: initialLoadingNotifier,
-        builder: (_, loading, __) {
-          if (loading) return const Center(child: CircularProgressIndicator());
-
-          return ValueListenableBuilder(
-            valueListenable: notifier,
-            builder: (_, state, __) {
-              return SliverScaffold(
-                appBar: SliverAppBarMedium(
-                  title: Text(strings.creditCard),
-                  loading: notifier.isLoading,
-                  actions: [
-                    FilledButton(
-                      onPressed: save,
-                      child: Text(strings.save),
-                    ),
-                    if (!state.creditCard.isNew)
-                      IconButton(
-                        tooltip: strings.delete,
-                        onPressed: delete,
-                        icon: const Icon(Icons.delete_outline),
-                      ),
-                  ],
-                ),
-                body: Form(
-                  key: formKey,
-                  child: Column(
-                    children: [
-                      const Spacing.y(),
-                      Padding(
-                        padding: const EdgeInsets.symmetric(horizontal: 16),
-                        child: Column(
-                          children: [
-                            TextFormField(
-                              initialValue: state.creditCard.description,
-                              decoration: InputDecoration(label: Text(strings.description)),
-                              validator: InputRequiredValidator().validate,
-                              textInputAction: TextInputAction.next,
-                              enabled: !notifier.isLoading,
-                              onSaved: (String? value) => state.creditCard.description = value ?? '',
-                            ),
-                            const Spacing.y(),
-                            TextFormField(
-                              initialValue: state.creditCard.amountLimit.moneyWithoutSymbol,
-                              decoration: InputDecoration(
-                                label: Text(strings.limit),
-                                prefixText: NumberFormat.simpleCurrency(locale: R.locale.toString()).currencySymbol,
-                              ),
-                              validator: InputRequiredValidator().validate,
-                              keyboardType: TextInputType.number,
-                              textInputAction: TextInputAction.next,
-                              enabled: !notifier.isLoading,
-                              onSaved: (String? value) => state.creditCard.amountLimit = (value ?? '').moneyToDouble(),
-                              inputFormatters: [FilteringTextInputFormatter.digitsOnly, MaskInputFormatter.currency()],
-                            ),
-                            const Spacing.y(),
-                            Row(
-                              children: [
-                                DropdownMenu<int>(
-                                  initialSelection: state.creditCard.invoiceClosingDay,
-                                  label: Text(strings.closingDay),
-                                  dropdownMenuEntries: List.generate(31, (index) => DropdownMenuEntry(value: index + 1, label: (index + 1).toString())),
-                                  onSelected: (int? value) => state.creditCard.invoiceClosingDay = value ?? 0,
-                                ),
-                                const Spacing.x(),
-                                DropdownMenu<int>(
-                                  initialSelection: state.creditCard.invoiceDueDay,
-                                  label: Text(strings.dueDay),
-                                  dropdownMenuEntries: List.generate(31, (index) => DropdownMenuEntry(value: index + 1, label: (index + 1).toString())),
-                                  onSelected: (int? value) => state.creditCard.invoiceDueDay = value ?? 0,
-                                ),
-                              ],
-                            ),
-                          ],
-                        ),
-                      ),
-                      const Spacing.y(),
-                      const Divider(),
-                      GroupTile(
-                        title: strings.brand,
-                        onTap: selectBrand,
-                        enabled: !notifier.isLoading,
-                        tile: state.creditCard.brand == null
-                            ? ListTile(
-                                leading: const Icon(Icons.style_outlined),
-                                title: Text(strings.selectBrand),
-                                trailing: const Icon(Icons.chevron_right),
-                                enabled: !notifier.isLoading,
-                              )
-                            : ListTile(
-                                leading: state.creditCard.brand?.icon(),
-                                title: Text(state.creditCard.brand?.description ?? ''),
-                                trailing: const Icon(Icons.chevron_right),
-                                enabled: !notifier.isLoading,
-                              ),
-                      ),
-                      const Divider(),
-                      GroupTile(
-                        onTap: selectAccount,
-                        title: strings.account,
-                        enabled: !notifier.isLoading,
-                        tile: state.creditCard.idAccount != null
-                            ? Builder(
-                                builder: (_) {
-                                  final AccountEntity account = accountsNotifier.value.accounts.firstWhere((account) => account.id == state.creditCard.idAccount);
-                                  return ListTile(
-                                    leading: account.financialInstitution!.icon(),
-                                    title: Text(account.description),
-                                    trailing: const Icon(Icons.chevron_right),
-                                    enabled: !notifier.isLoading,
-                                  );
-                                },
-                              )
-                            : ListTile(
-                                leading: const Icon(Icons.account_balance_outlined),
-                                title: Text(strings.selectAccount),
-                                trailing: const Icon(Icons.chevron_right),
-                                enabled: !notifier.isLoading,
-                              ),
-                      ),
-                      const Divider(),
-                    ],
+      valueListenable: initialLoadingNotifier,
+      builder: (_, initialLoading, __) {
+        return ValueListenableBuilder(
+          valueListenable: notifier,
+          builder: (_, state, __) {
+            return SliverScaffold(
+              appBar: SliverAppBarMedium(
+                title: Text(strings.creditCard),
+                loading: notifier.isLoading,
+                actions: [
+                  FilledButton(
+                    onPressed: save,
+                    child: Text(strings.save),
                   ),
-                ),
-              );
-            },
-          );
-        });
+                  if (!state.creditCard.isNew)
+                    IconButton(
+                      tooltip: strings.delete,
+                      onPressed: delete,
+                      icon: const Icon(Icons.delete_outline),
+                    ),
+                ],
+              ),
+              body: Builder(
+                builder: (_) {
+                  if (initialLoading) return const SizedBox.shrink();
+
+                  return Form(
+                    key: formKey,
+                    child: Column(
+                      children: [
+                        const Spacing.y(),
+                        Padding(
+                          padding: const EdgeInsets.symmetric(horizontal: 16),
+                          child: Column(
+                            children: [
+                              TextFormField(
+                                initialValue: state.creditCard.description,
+                                decoration: InputDecoration(label: Text(strings.description)),
+                                validator: InputRequiredValidator().validate,
+                                textInputAction: TextInputAction.next,
+                                enabled: !notifier.isLoading,
+                                onSaved: (String? value) => state.creditCard.description = value ?? '',
+                              ),
+                              const Spacing.y(),
+                              TextFormField(
+                                initialValue: state.creditCard.amountLimit.moneyWithoutSymbol,
+                                decoration: InputDecoration(
+                                  label: Text(strings.limit),
+                                  prefixText: NumberFormat.simpleCurrency(locale: R.locale.toString()).currencySymbol,
+                                ),
+                                validator: InputRequiredValidator().validate,
+                                keyboardType: TextInputType.number,
+                                textInputAction: TextInputAction.next,
+                                enabled: !notifier.isLoading,
+                                onSaved: (String? value) => state.creditCard.amountLimit = (value ?? '').moneyToDouble(),
+                                inputFormatters: [FilteringTextInputFormatter.digitsOnly, MaskInputFormatter.currency()],
+                              ),
+                              const Spacing.y(),
+                              Row(
+                                children: [
+                                  DropdownMenu<int>(
+                                    initialSelection: state.creditCard.invoiceClosingDay,
+                                    label: Text(strings.closingDay),
+                                    dropdownMenuEntries: List.generate(31, (index) => DropdownMenuEntry(value: index + 1, label: (index + 1).toString())),
+                                    onSelected: (int? value) => state.creditCard.invoiceClosingDay = value ?? 0,
+                                  ),
+                                  const Spacing.x(),
+                                  DropdownMenu<int>(
+                                    initialSelection: state.creditCard.invoiceDueDay,
+                                    label: Text(strings.dueDay),
+                                    dropdownMenuEntries: List.generate(31, (index) => DropdownMenuEntry(value: index + 1, label: (index + 1).toString())),
+                                    onSelected: (int? value) => state.creditCard.invoiceDueDay = value ?? 0,
+                                  ),
+                                ],
+                              ),
+                            ],
+                          ),
+                        ),
+                        const Spacing.y(),
+                        const Divider(),
+                        GroupTile(
+                          title: strings.brand,
+                          onTap: selectBrand,
+                          enabled: !notifier.isLoading,
+                          tile: state.creditCard.brand == null
+                              ? ListTile(
+                                  leading: const Icon(Icons.style_outlined),
+                                  title: Text(strings.selectBrand),
+                                  trailing: const Icon(Icons.chevron_right),
+                                  enabled: !notifier.isLoading,
+                                )
+                              : ListTile(
+                                  leading: state.creditCard.brand?.icon(),
+                                  title: Text(state.creditCard.brand?.description ?? ''),
+                                  trailing: const Icon(Icons.chevron_right),
+                                  enabled: !notifier.isLoading,
+                                ),
+                        ),
+                        const Divider(),
+                        GroupTile(
+                          onTap: selectAccount,
+                          title: strings.account,
+                          enabled: !notifier.isLoading,
+                          tile: state.creditCard.idAccount != null
+                              ? Builder(
+                                  builder: (_) {
+                                    final AccountEntity account = accountsNotifier.value.accounts.firstWhere((account) => account.id == state.creditCard.idAccount);
+                                    return ListTile(
+                                      leading: account.financialInstitution!.icon(),
+                                      title: Text(account.description),
+                                      trailing: const Icon(Icons.chevron_right),
+                                      enabled: !notifier.isLoading,
+                                    );
+                                  },
+                                )
+                              : ListTile(
+                                  leading: const Icon(Icons.account_balance_outlined),
+                                  title: Text(strings.selectAccount),
+                                  trailing: const Icon(Icons.chevron_right),
+                                  enabled: !notifier.isLoading,
+                                ),
+                        ),
+                        const Divider(),
+                      ],
+                    ),
+                  );
+                },
+              ),
+            );
+          },
+        );
+      },
+    );
   }
 
   Future<void> save() async {
@@ -204,14 +210,13 @@ class _CreditCardFormPageState extends State<CreditCardFormPage> with ThemeConte
         formKey.currentState?.save();
 
         await notifier.save();
+        if (notifier.value is ErrorCreditCardState) throw Exception((notifier.value as ErrorCreditCardState).message);
 
         if (!mounted) return;
         context.pop(FormResultNavigation.save(notifier.creditCard));
       }
     } catch (e) {
-      ErrorDialog.show(context, e.toString());
-    } finally {
-      notifier.value = notifier.value.changedCreditCard();
+      await ErrorDialog.show(context, e.toString());
     }
   }
 
@@ -220,13 +225,12 @@ class _CreditCardFormPageState extends State<CreditCardFormPage> with ThemeConte
 
     try {
       await notifier.delete();
+      if (notifier.value is ErrorCreditCardState) throw Exception((notifier.value as ErrorCreditCardState).message);
 
       if (!mounted) return;
       context.pop(FormResultNavigation<CreditCardEntity>.delete());
     } catch (e) {
-      ErrorDialog.show(context, e.toString());
-    } finally {
-      notifier.value = notifier.value.changedCreditCard();
+      await ErrorDialog.show(context, e.toString());
     }
   }
